refactor: 이미지 업로드 api 파일경로 변경, 아이콘 컴포넌트화,
