+++ src/api/uploadProfile.js
@@ -1,4 +1,4 @@
-export async function uploadImage(file) {
+export async function uploadProfileImage(file) {
   const data = new FormData();
   data.append('file', file);
   data.append('upload_preset', process.env.REACT_APP_CLOUD_API_PROFILE);
