+++ src/components/common/iconComponent/IconComponents.jsx
@@ -254,22 +254,6 @@ const icons = {
       />
     </>
   ),
-  signupCheck: (
-    <>
-      <g id='Group 16'>
-        <g id='Group 15'>
-          <circle id='Ellipse 17' cx='36' cy='36' r='36' fill='#5873FE' />
-        </g>
-        <path
-          id='Vector'
-          d='M18.5 37.5L29.0128 48.998C29.2494 49.3055 29.5526 49.5553 29.8997 49.7288C30.2467 49.9023 30.6285 49.995 31.0165 49.9998C31.3982 50.0042 31.7761 49.9235 32.1227 49.7636C32.4692 49.6036 32.7758 49.3684 33.0201 49.0751L55.3943 22'
-          strokeWidth='4.45257'
-          strokeLinecap='round'
-          strokeLinejoin='round'
-        />
-      </g>
-    </>
-  ),
   pencil: (
     <>
       <g id='&#235;&#139;&#137;&#235;&#132;&#164;&#236;&#158;&#132; &#235;&#176;&#148;&#234;&#190;&#184;&#234;&#184;&#176; Icon'>
