--- src/layout/header/PostMainHeader.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function PostPageHeader() {
-  return (
-    <div>PostPageHeader</div>
-  )
-}
-
-export default PostPageHeader
