+++ src/pages/group/GroupEdit.jsx
@@ -2,12 +2,13 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
 import api from '../../api/index.jsx';
-import { uploadImage } from '../../hooks/uploadImage.js';
+import { uploadImage } from '../../api/uploadImage.js';
 import WriteImageUpload from '../../components/common/input/WriteImageUpload.jsx';
 import { DatePicker, Space } from 'antd';
 import Input from '../../components/common/input/Input.jsx';
 import moment from 'moment';
 import FriendSearchModal from '../../components/common/modal/NicknameModal.jsx';
+import IconComponents from '../../components/common/iconComponent/IconComponents.jsx';
 
 
 function GroupWrite() {
@@ -184,10 +185,10 @@ function GroupWrite() {
       <Form onSubmit={submitHandler}>
         <WriteHeader>
           <div>
-            <BackButton onClick={backButtonHandler}>
-              <img
-                src={`${process.env.PUBLIC_URL}/assets/svgs/icon_back.svg`}
-                alt='left'
+          <BackButton onClick={backButtonHandler}>
+              <IconComponents
+                iconType='vectorLeft'
+                stroke ='#4C4C4C'
               />
             </BackButton>
           </div>
@@ -442,6 +443,8 @@ const BackButton = styled.button`
   margin-left: 20px;
   background-color: transparent;
   border: none;
+  position:relative;
+  top:5px
 `;
 
 const WriteImageWrapper = styled.div`
